+++ app/forms/webhooks/create_form.rb
@@ -27,13 +27,20 @@ module Webhooks
       'There is already webhook with such source'
     end
 
+    # rubocop: disable Metrics/CyclomaticComplexity
     def validate_url(params)
       case params[:source]
       when Webhook::SLACK
         params[:url].start_with?('https://hooks.slack.com/services') ? nil : 'Invalid Slack webhook url'
       when Webhook::DISCORD
         params[:url].start_with?('https://discord.com/api/webhooks') ? nil : 'Invalid Discord webhook url'
+      when Webhook::CUSTOM
+        nil if URI(params[:url]).origin
       end
+    # catches undefined method `origin' for #<URI::Generic> if url is invalid
+    rescue NoMethodError => _e
+      'Invalid Webhook url'
     end
+    # rubocop: enable Metrics/CyclomaticComplexity
   end
 end